backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==3.7.1
astrapy==2.0.1
asttokens==3.0.0
attrs==25.3.0
beanie==1.30.0
cachetools==5.5.2
cassandra-driver==3.29.2
cassio==0.1.10
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
comm==0.2.2
cryptography==45.0.4
dataclasses-json==0.6.7
debugpy==1.8.14
decorator==5.2.1
Deprecated==1.2.18
deprecation==2.1.0
dnspython==2.7.0
email_validator==2.2.0
executing==2.2.0
fastapi[standard]==0.115.13
fastapi-cli==0.0.7
filelock==3.18.0
filetype==1.2.0
frozenlist==1.7.0
fsspec==2025.5.1
geomet==0.2.1.post1
google-ai-generativelanguage==0.6.18
google-api-core==2.25.1
google-auth==2.40.3
google-genai==1.20.0
google-generativeai==0.3.2
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.73.0
grpcio-status==1.62.3
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.33.0
hyperframe==6.1.0
idna==3.10
ipykernel==6.29.5
ipython==9.3.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_widgets==3.0.15
langchain==0.3.25
langchain-astradb==0.6.0
langchain-community==0.3.25
langchain-core==0.3.65
langchain-google-genai==2.1.5
langchain-huggingface==0.3.0
langchain-text-splitters==0.3.8
langsmith==0.3.45
lazy-model==0.2.0
lxml==5.4.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mdurl==0.1.2
motor==3.7.1
mpmath==1.3.0
multidict==6.4.4
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.5
nltk==3.8.1
numpy==1.26.4
orjson==3.10.18
packaging==24.2
parso==0.8.4
pillow==11.2.1
platformdirs==4.3.8
prompt_toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==4.25.8
psutil==7.0.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
PyJWT==2.10.1
pymongo==4.13.1
pypdf==5.6.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.1.0
python-dotenv==1.0.0
python-multipart==0.0.20
pywin32==310
PyYAML==6.0.2
pyzmq==27.0.0
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.0.0
rich-toolkit==0.14.7
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
sentence-transformers==4.1.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
stack-data==0.6.3
svix==1.67.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.21.1
toml==0.10.2
torch==2.7.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.52.4
typer==0.16.0
types-Deprecated==1.2.15.20250304
types-python-dateutil==2.9.0.20250516
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
uuid6==2025.0.0
uvicorn==0.34.3
watchfiles==1.0.5
wcwidth==0.2.13
websockets==15.0.1
widgetsnbextension==4.0.14
wrapt==1.17.2
yarl==1.20.1
zstandard==0.23.0
